Rename initializers folder by helpers folder && Update import at AddBook component
