+++ src/redux/helpers/categoryInitializer.js
